+++ apps/auth/node-and-express-server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # crossmint-auth-node
 
+## 1.1.79
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.68
+
+## 1.1.78
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.67
+
 ## 1.1.77
 
 ### Patch Changes