+++ apps/payments/nextjs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @crossmint/client-sdk-nextjs-starter
 
+## 2.0.18
+
+### Patch Changes
+
+- Updated dependencies [79d0e8f]
+  - @crossmint/client-sdk-base@2.3.1
+  - @crossmint/client-sdk-react-ui@4.1.2
+
+## 2.0.17
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@4.1.1
+
+## 2.0.16
+
+### Patch Changes
+
+- Updated dependencies [11f2089]
+  - @crossmint/client-sdk-react-ui@4.1.0
+  - @crossmint/client-sdk-base@2.3.0
+
 ## 2.0.15
 
 ### Patch Changes