+++ apps/wallets/quickstart-devkit/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @crossmint/wallets-quickstart-devkit
 
+## 0.2.38
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@4.1.2
+
+## 0.2.37
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@4.1.1
+
+## 0.2.36
+
+### Patch Changes
+
+- Updated dependencies [11f2089]
+  - @crossmint/client-sdk-react-ui@4.1.0
+
 ## 0.2.35
 
 ### Patch Changes