+++ packages/client/verifiable-credentials/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @crossmint/client-sdk-verifiable-credentials
 
+## 3.4.84
+
+### Patch Changes
+
+- Updated dependencies [79d0e8f]
+  - @crossmint/client-sdk-base@2.3.1
+
+## 3.4.83
+
+### Patch Changes
+
+- Updated dependencies [11f2089]
+  - @crossmint/client-sdk-base@2.3.0
+
 ## 3.4.82
 
 ### Patch Changes