+++ packages/server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/server-sdk
 
+## 1.2.68
+
+### Patch Changes
+
+- @crossmint/common-sdk-auth@1.1.5
+
+## 1.2.67
+
+### Patch Changes
+
+- @crossmint/common-sdk-auth@1.1.4
+
 ## 1.2.66
 
 ### Patch Changes